Declare factorization flags in sprs rather than sprs-ldl
The rationale is that other crates might depend on this info.
